Explicitly set the visibility of right_icon within call notification
The right icon within the notification at times had the wrong visibility
state. This led to the contact attribution badge not showing up at times.
This could be the result of a series of notification state updates due to
the state change callbacks via ContactInfoCache. Resorting to explicitly
updating the view's visibility.

Change-Id: I447a1a4dc08ee8b6bba5569c7a1bb91d7677bd15
Issue-Id: OPO-766